pacifica-edge/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
boto3>=1.40.0,<2
email-validator>=2.2.0,<3
